--- openstack/blockstorage/extensions/services/requests.go
@@ -1,45 +1,0 @@
-package services
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
-)
-
-// ListOptsBuilder allows extensions to add additional parameters to the List
-// request.
-type ListOptsBuilder interface {
-	ToServiceListQuery() (string, error)
-}
-
-// ListOpts holds options for listing Services.
-type ListOpts struct {
-	// Filter the service list result by binary name of the service.
-	Binary string `q:"binary"`
-
-	// Filter the service list result by host name of the service.
-	Host string `q:"host"`
-}
-
-// ToServiceListQuery formats a ListOpts into a query string.
-func (opts ListOpts) ToServiceListQuery() (string, error) {
-	q, err := golangsdk.BuildQueryString(opts)
-	if err != nil {
-		return "", err
-	}
-	return q.String(), err
-}
-
-// List makes a request against the API to list services.
-func List(client *golangsdk.ServiceClient, opts ListOptsBuilder) pagination.Pager {
-	url := listURL(client)
-	if opts != nil {
-		query, err := opts.ToServiceListQuery()
-		if err != nil {
-			return pagination.Pager{Err: err}
-		}
-		url += query
-	}
-	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
-		return ServicePage{pagination.SinglePageBase(r)}
-	})
-}
