--- openstack/blockstorage/extensions/services/urls.go
@@ -1,7 +1,0 @@
-package services
-
-import "github.com/opentelekomcloud/gophertelekomcloud"
-
-func listURL(c *golangsdk.ServiceClient) string {
-	return c.ServiceURL("os-services")
-}
